Convert updated at timestamp to local timezone
Prior to this change, this was showing the times one hour out during BST
(because we're an hour ahead of UTC, so everything was '1 hour ago' or
longer).

This change converts the time to the local timezone. We then need to
strip the timezone info out to allow it to be compared to the current
time in ``humanize``, which can only handle naive datetimes.